+++ app/[lang]/(app)/(home)/(tablet)/room-devision/execution-management/daily-tasks-checklist/utils/dailyTasksChecklistProps.ts
@@ -1,0 +1,9 @@
+import { type CheckList } from '../services/dailyTasksApiActions';
+
+export interface DailyTasksChecklistProps {
+ data?: CheckList[];
+ isFetching: boolean;
+ refetch: () => unknown;
+ isSuccess: boolean;
+ isError: boolean;
+}