+++ app/[lang]/(app)/(home)/(tablet)/room-devision/execution-management/daily-tasks-checklist/utils/editDailyTaskProps.ts
@@ -1,0 +1,9 @@
+import { type CheckList } from '../services/dailyTasksApiActions';
+
+export interface EditDailyTaskProps {
+ selectedCheckListID: null | number;
+ showNew: boolean;
+ onShowEdit: (id: number | null) => unknown;
+ onCloseEdit: () => unknown;
+ targetEditChecklist: null | CheckList;
+}