+++ app/[lang]/(app)/(home)/(tablet)/room-devision/execution-management/daily-tasks-checklist/utils/timeNo.ts
@@ -1,0 +1,9 @@
+const timeNoOptions = ['firstTurn', 'secondTurn', 'thirdTurn'] as const;
+
+enum TimeNo {
+ firstTurn = 1,
+ secondTurn,
+ thirdTurn,
+}
+
+export { timeNoOptions, TimeNo };